requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
asgiref==3.8.1
async-timeout==5.0.1
asyncio==3.4.3
attrs==25.3.0
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.4.26
charset-normalizer==3.4.2
Django==4.2.21
django-cors-headers==4.7.0
django-filter==25.1
django-taggit==6.1.0
djangorestframework==3.16.0
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.5.1
gensim==4.3.3
hf-xet==1.1.5
huggingface-hub==0.33.2
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
Markdown==3.8.2
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.4.4
networkx==3.5
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
packaging==25.0
pgvector==0.4.1
pillow==11.3.0
propcache==0.3.1
psycopg2==2.9.10
pypdf==5.7.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.13.1
sentence-transformers==5.0.0
smart-open==7.1.0
soupsieve==2.7
sqlparse==0.5.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
transformers==4.53.1
triton==3.3.1
typing_extensions==4.13.2
urllib3==2.4.0
wrapt==1.17.2
yarl==1.20.0
